Add gen4_state_t structure
The idea here is to actually use a (padded) C structure instead
of just a bunch of void* addition with global offset values and
casts.

So far, we only have the kernels in the structure, (and we're even
doing more copying of kernels than before). But soon enough we'll
switch to separate "general" and "surface" state structures which
will let us avoid do any kernel copying whatsover while compositing.