Remove spurious use of MI_SET_CONTEXT.
The driver uses MI_SET_CONTEXT to save the current state before
emitting the invariant 3d state.  Nothing ever restores this state though,
so this patch removes that.  This was also the only remaining use of
the ring buffer in TTM mode, so we're one step closer to a ring-buffer
free DDX driver.