validate buffers into TTM space only.
If we validate into both VRAM and TTM here, then later one we emit a relocation
to only one of the memory types (i.e. TTM) the memory manager seems to lose
the plot and keep allocating memory blocks in the TTM space even though the
buffer is already there.