textured xv: Allocate state buffer per frame instead of reusing it.
Relocations are never cleared since batch/state buffers aren't supposed
to be reused.  This leads to a crash after approximately 30 seconds of
video when we run out of relocs.

This fixes the textured video to just allocated the state buffer on the fly,
which is also a nice cleanup and avoid unnecessary syncing.