Force unmap and mark BOs as shared when we get the underlying handle.
The bufmgr optimizes mapping and unmapping for unshared buffer objects.
When it's the only user of a buffer object, it delays the actual
drmBOUnmap call until absolutely necessary, which is when we need to
submit it.  So when we pass the drm bo handle to the DRI driver, we
need to force unmap it and mark it shared.