Poll DP link every 10s and retrain as needed.
If the DP cable is removed, the link goes down and must be retrained when it
is plugged back in. In a kernel mode driver, this would be done with hotplug
detection, but up in user mode, we can't easily do that. Instead, just poll
every 10 seconds and retrain when the connector is plugged back in.

Signed-off-by: Keith Packard <keithp@keithp.com>