Just add all of the GPIO-detected digital outputs
We can't distinguish between HDMI and DP outputs as they
share the same GPIO pin. The only way to tell is to parse the
BIOS tables. Instead of doing that, just add all of the outputs and let
monitor detection sort it out (which is reliable on DP and HDMI).

Signed-off-by: Keith Packard <keithp@keithp.com>