Save/restore DP link/gmch ratios. Set link ratios before enabling pipe.
The DP link ratios control how the pipe sends data to the output, and so
should be configured before the pipe is turned on. This should make the
output work the first time it is enabled. This bug was masked because the
server didn't save and restore the ratio registers, so subsequent calls to
enable the output would work using the values previously stored in those
registers.

Signed-off-by: Keith Packard <keithp@keithp.com>